+++ js/index.ts
@@ -720,6 +720,14 @@ const resetTrackingConfig = (successCallback: any, errorCallback: any) => {
   NativeModules.RNRoam.resetTrackingConfig(successCallback, errorCallback);
 };
 
+const checkActivityPermission = (callback: any) => {
+  NativeModules.RNRoam.checkActivityPermission(callback);
+};
+
+const requestActivityPermission = () => {
+  NativeModules.RNRoam.requestActivityPermission();
+};
+
 const Roam = {
   TrackingMode,
   DesiredAccuracy,
@@ -800,6 +808,8 @@ const Roam = {
   RoamTrip,
   RoamTripStop,
   RoamCustomTrackingOptions,
+  checkActivityPermission,
+  requestActivityPermission,
 };
 
 export default Roam;